+++ test/support/echo.ex
@@ -1,0 +1,3 @@
+defmodule Vor.TestHelpers.Echo do
+  def reflect(value), do: value
+end